Refactor ProfileScreen into multiple classes, and use the hiltviewModel
