+++ app/src/main/java/ch/epfl/cs311/wanderwave/viewmodel/ProfileViewModel.kt
@@ -2,10 +2,13 @@ package ch.epfl.cs311.wanderwave.viewmodel
 
 import androidx.lifecycle.ViewModel
 import ch.epfl.cs311.wanderwave.model.data.Profile
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import javax.inject.Inject
 
-class ProfileViewModel : ViewModel() {
+@HiltViewModel
+class ProfileViewModel @Inject constructor() : ViewModel(){
   private val _profile =
       MutableStateFlow(
           Profile(
@@ -34,4 +37,5 @@ class ProfileViewModel : ViewModel() {
   fun togglePublicMode() {
     _isInPublicMode.value = !_isInPublicMode.value
   }
+
 }
